+++ R/package-SHT.R
@@ -11,7 +11,7 @@
 #' @noRd
 #' @aliases SHT-package
 #' @import Rdpack
-#' @importFrom flare sugm
+#' @import flare
 #' @importFrom pracma pinv strcmp
 #' @importFrom utils packageVersion
 #' @importFrom stats pt sd cov pchisq pf median aov cov2cor pnorm var runif shapiro.test rgamma rlnorm cor ppoints qnorm dist rnorm qnorm punif uniroot